video: tegra: host: fix 64 bit regops
In regops, there is a whitelist of allowed registers. The
whitelisted registers are always considered to be 32-bit.
However, we do not consider 64-bit read/write operations
during whitelist validation.
e.g. if register at 0x400500 is whitelisted, with a 64-bit
read we can also read register at 0x400504 which is not
whitelisted

To fix this, first separate out check_whitelists() function
from validate_reg_op_offset() which does the whitelist
validation of given offset.
Then pass offset to this function to check if it is
whitelisted or not.
If yes and if operation is 64-bit, then pass (offset + 4)
to check_whitelists() to check if next offset is also
whitelisted or not.

Bug 1440107

Change-Id: I9a08cc031e741987f5577d47d32b303ba114ff07
Signed-off-by: Deepak Nibade <dnibade@nvidia.com>
Reviewed-on: http://git-master/r/368437
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Tested-by: Terje Bergstrom <tbergstrom@nvidia.com>