net: wireless: bcmdhd: fix 64-bit safeness issues
If a structure has pointer as a field and used in copy_from_user,
it causes issues in case of 32-bit userspace and 64-bit kernel
as structure size varies in kernel and userspace. To handle this case
define new structure without a pointer and use in case of compat task.

bug 1466409

Change-Id: I1dfb511b2683d93ba66626877885a4ad0ea37470
Signed-off-by: Manikanta <mmaddireddy@nvidia.com>
Reviewed-on: http://git-master/r/374795
Reviewed-by: Nagarjuna Kristam <nkristam@nvidia.com>
Tested-by: Nagarjuna Kristam <nkristam@nvidia.com>
Reviewed-by: Venu Byravarasu <vbyravarasu@nvidia.com>