ARM64: tegra: add empty BoardConfig.mk
For android-3.10 merge to dev-kernel, we are setting a flag
in BoardConfig.mk and hence we need to have a clean
build. But with modular builds, only kernel is compiled.
This breaks Norrin boot.

To fix this, add empty BoardConfig.mk file with this change.
Once we have Android build with this change, any change to this
file will trigger Android build.
And after that we can set the flag in BoardConfig.mk

Bug 1456092

Change-Id: Ie32b6847862356139927182da9babd86bbdef4db
Signed-off-by: Deepak Nibade <dnibade@nvidia.com>
Reviewed-on: http://git-master/r/377707
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>