phy: core: Add an exported of_phy_get function
Previously the of_phy_get function took a struct device * and
was declared static. It was impossible to call it from
another driver and thus it was impossible to get phy defined
for a given node. The old function was renamed to _of_phy_get
and was left for internal use. of_phy_get function was added
and it was exported. The function enables to get a phy for
a given device tree node.

Signed-off-by: Kamil Debski <k.debski@samsung.com>
Reviewed-by: Tomasz Figa <t.figa@samsung.com>
Signed-off-by: Kishon Vijay Abraham I <kishon@ti.com>
(cherry picked from commit 0b3f3b2c777a2f7d20c9826a190ffd5bbd288f8f)

Bug 200024489

Signed-off-by: Rakesh Babu Bodla <rbodla@nvidia.com>
Change-Id: Ib513e84ed5d55d593f81b5712fa3a29427086966
Reviewed-on: http://git-master/r/682490
GVS: Gerrit_Virtual_Submit
Reviewed-by: Venkat Moganty <vmoganty@nvidia.com>