thermal: core: use cancel delayed work sync
This fixes an issue with race between a piece of delayed work in
progress and its cancellation during sensor shutdown that happens as
part of system power-down. We refactor the code a bit to make sure the
cancel_sync is called outside the tz mutex to prevent deadlock.

This was found during a stress test by continuously running warm-reboot
in a loop. We found that eventually, we get a system hang or crash with
an error message from a failed reading of the temp from the get_temp API
which indicates a potential issue with the polling work function.
We can accelerate the reproduction of the issue by introducing an
artificial delay in the work function and forcing polling by setting a
non-zero polling_delay for the thermal zone.

In the actual system, the timing differences could be introduced from
other unresolved errors in the I2C bus transactions which could show
up during the cold and warm reboot stress tests.

Bug 1463497

Change-Id: I58b8a90052850b0d67d531200d90ed2dc168e48c
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Cc: Paul Walmsley <pwalmsley@nvidia.com>
Cc: Seema Khowala <seemaj@nvidia.com>
Reviewed-on: http://git-master/r/378692
Reviewed-by: Aleksandr Frid <afrid@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Seema Khowala <seemaj@nvidia.com>
Tested-by: Seema Khowala <seemaj@nvidia.com>