drivers: gadget: tegra: support to skip charging
Currently charging_support flag is left unimplemented in the UDC driver and it
is assumed that all platforms need/support charging.  Add support to make this
optional so that platforms can choose to skip charger port detection and
assume instead that it is always connected to only SDP ports.

Bug 1463801

Change-Id: I89b27e9a8d5af4d29af0609bfc5faa643d2d990c
Signed-off-by: Haribabu Narayanan <hnarayanan@nvidia.com>
Reviewed-on: http://git-master/r/378732
Reviewed-by: Venkat Moganty <vmoganty@nvidia.com>