iommu/arm-smmu: allow 42-bit stage-1 output size with 64K pages
The output size of stage-1 is currently limited by the input size of
stage-2, which is further limited by VA_BITS since we make use of the
standard pgd_alloc functions for creating page tables.

This patch ensures that we use VA_BITS instead of a hardcoded '39'
for the stage-1 output size limit.

Change-Id: Ia967696977e12ddad8fa9f22a3da46666d06a933
Signed-off-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Hiroshi Doyu <hdoyu@nvidia.com>