arm: tegra: soctherm: fix pointer misuse in zoneid
Storing the integer local thermal array index in tz->devdata pointer
is a misuse of pointers and gives rise to compilation warnings in
64bit builds.

Fix this by storing the pointer to the local therm structure and
retrieving it in the called routines. Since we can no longer have a
thermal zone for each tsensor, we don't support it anymore.

When handling soctherm HW interrupt, we know the exact trip_point
triggered. There's no need to update the entire zone. Hence we switch
to using the thermal_notify_framework API that takes the
specific trip_point.

Bug 1478989

Change-Id: I2211114602d54a4b0ca9d9780ce9a0b68e1ccecc
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/385817