arm64: optimize exception vector branch prediction
A common static branch predictor in hardware is to keep track of the
branch stack in hardware so that RET will always predict the correct
return address. Manually setting the link register with something like
the following will result in branch misprediction on the RET

  adr lr, bar
  b   foo

A branch misprediction throws away energy and performance by executing
the mispredicted path.

The system call handler, el0_svc_naked, does exactly that.
This patch switches to using blr and adds an extra branch to trade off
for the branch misprediction.

The synchronous exception handler, el0_sync_compat, does that too.
This patch switches to using blr, adds a few more instructions, and adds
an extra branch to trade off for the branch misprediction.

Bug 1488764

Change-Id: If1c40c432ab8822fbca21f5618353a6aad1b7ec6
Signed-off-by: Alex Van Brunt <avanbrunt@nvidia.com>
Reviewed-on: http://git-master/r/386536
GVS: Gerrit_Virtual_Submit
Reviewed-by: Richard Wiley <rwiley@nvidia.com>
Reviewed-by: Peng Du <pdu@nvidia.com>