arm: tegra13: soctherm: restore orig threshold
For certain older revs of ATE, we will use use PLL-TSOSC (instead of
the CPU-TSOSCs) on T132 devices to drive throttling (HW and SW) and
shutdown to avoid random shutdown issues seen on some platforms.

Revert T132-specific hacks in soctherm driver to prepare for the
change, which is done in per-platform board-files.

Bug 1468124

Change-Id: Iac8152045915ac59db45c9b30103dbbb13a784b2
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/395413