video: tegra: nvmap: track cpu dirty pages
A page is marked cpu dirty when a user space access faults or
when kernel access happens through dma_buf_kmap.
A page is marked cpu clean when it is written back. This operation is valid
only for write back i.e., when buffer is sync for device.
For kernel accesses through kmap/vmap it is assumed that client calls
begin/end_cpu_access to take care of sync for device/cpu immediately after
access and hence we do not track such dirty pages.
enable tracking cpu dirty pages, if auto_cache_sync is enabled for handle.

Bug 1444151

Change-Id: Ic03478702fad67b6d7682eb20404e3f88114d81c
Signed-off-by: Sri Krishna chowdary <schowdary@nvidia.com>
Signed-off-by: Krishna Reddy <vdumpa@nvidia.com>
Reviewed-on: http://git-master/r/391816
Reviewed-by: Automatic_Commit_Validation_User