irqchip: gic: no error print if cpu number is 1
if GIC is configured for a uniprocessor, the TARGETSRs will be read
as 0, so cpumask will be always 0. do not print error message in such
case. if GIC is configured for a multiprocessor, the TARGETSRs will
be read back as non-0, then it makes sense to print an error if
cpumask is 0. however, this change can't handle the case when
there is only 1 CPU in a multiprocessor configuration, in that case
cpumask shall not be 0, but the number of CPUs is 1.

bug 1500806

Change-Id: I0da7b763f9374e77e4a2d1d6c6b7c8299f09fb3b
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: http://git-master/r/397371
Tested-by: Arun Swain <arswain@nvidia.com>
Reviewed-by: Automatic_Commit_Validation_User