arm: tegra13: Always enfore 2-core CPU EDP limit
On T132, offlining a core within Linux is not sufficient to guarantee that
the core is continually in a powered-down state. Until we do have a sufficient
guarantee, we will use EDP tables to always enforce the 2-core freq cap even
when a single core is ONLINE.

To do this in the simplest way, values in the dynamic_constants[] array and
the leakage_constants[] array are all set to the 2-core value below in the
structure init. With this, there's no need to modify the code edp logic in
edp.c and it remains specific to this chip.

The actual 1-core values for these two arrays is saved as a comment to
resurrect when actually needed.

Bug 1434482
Bug 1427964

Change-Id: I295c21ae68b99626e824a1210479155ce9cbf279
Signed-off-by: Diwakar Tundlam <dtundlam@nvidia.com>
Reviewed-on: http://git-master/r/399106
Reviewed-by: Sai Gurrappadi <sgurrappadi@nvidia.com>