Asoc:tegra:rt5639: Fix false headset intr issue
When the device enters to LP0, resetting of the audio codec makes the
HP detection GPIO pin high and causes a false interrupt. If the codec
gets reset before the interrupt is disabled by the machine driver,
device never goes to LP0 or wakes up from LP0. Added fix to avoid
this condition.

Bug 1499603

Change-Id: I5fd498c13a384b501243b13a7faef6aa069aa7ab
Signed-off-by: Sharad Gupta <sharadg@nvidia.com>
Reviewed-on: http://git-master/r/400786
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sumit Bhattacharya <sumitb@nvidia.com>
Reviewed-by: Ravindra Lokhande <rlokhande@nvidia.com>
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>