IKJBXLINE-7047: f2fs: mask off 0007 on inodes marked as root
This commit allows the "other" permissions to be masked when an
inode has a special system.advise extended attribute set.  It
cleans up some redundancy in the android_emu checks and prepares
the way for the use of ACLs to accomplish this.

Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
(cherry picked from commit b34ca72121ae57098688cff1f920d34a6c88f0dd)

Change-Id: I16e4b245d57bad681ef885fe1e859463e61566b3