video: tegra: host: Fix race in sync dump
The code assumed that the sync framework will not call nvhost
specific functions until the pointer has been returned to
the caller. In most cases this is a fair assumption, however,
if the sync dump was in progress at the same time we created a
new fence, we may try to print out fence specific data before
it has been created.

This function adds necessary check to nvhost_sync_pt_value_str()
which is the only function suffering from this issue.

Bug 200001909

Change-Id: Ieaf8352eb48aedbb984e05caec241aca3e593810
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/407444
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>