ARM: Tegra: fix possible race in cluster switch
It's possible a core goes to C7 after BPMP has sent the active CPU mask. In
that case the kernel will still send an IPI, but the CPU won't wake up due to
the intercept set by BPMP. However the CPU will wake up after the cluster
switch, so we need to prevent it from going to forced C7 again. To do that the
existing code relies on the cluster_switch_pending variable. However this
variable is cleared by the last man in kernel space and the other CPUs can
already by active then as they are released by BPMP as soon as the last man
has set the ACTIVE_CLUSTER_LOCK bit in FLOW_CTLR_BPMP_CLUSTER_CONTROL. Instead
pass the expected cluster ID and use that to determine if the CPU has to go
to C7.

Change-Id: I8ada1513b2595cb3945c3705c2ed2d5ce40a0d1a
Signed-off-by: Peter De Schrijver <pdeschrijver@nvidia.com>
Reviewed-on: http://git-master/r/407635
Reviewed-by: Timo Alho <talho@nvidia.com>