pstore: use writecombine instead of noncached
pgprot_noncached implies that memory accesses cannot be cached and that
reads and writes cannot be combined. pgprot_writecombine implies that
accesses are non-cached but that reads and writes can be combined.
Because pstore is memory backed, the fact that reads and writes are
combined does not affect the behavior.

Bug 1512416

Change-Id: I9d1d651b146789d2ae8846ccddee72043b6198ee
Signed-off-by: Alex Van Brunt <avanbrunt@nvidia.com>
Reviewed-on: http://git-master/r/407802
Reviewed-by: Alex Waterman <alexw@nvidia.com>
Reviewed-by: Richard Wiley <rwiley@nvidia.com>
Tested-by: Richard Wiley <rwiley@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>