arm: tegra: headsmp: fix tegra_with_secure_firmware access with MMU off
There are places during the CPU resume path where we access this variable
with MMU off. In such scenarios we should use the physical address for this
variable.

This fixes the virtualisation team's issue, since they were the ones who
reported it in the first place. Fix a case where the code running from
iRAM was accessing the variable from DRAM instead of the one cached in
iRAM.

Bug 1411345

Change-Id: I9005c30329d38bae305a4a7b31ae7e2ca83e8a5d
Signed-off-by: Varun Wadekar <vwadekar@nvidia.com>
Reviewed-on: http://git-master/r/412540
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>
Reviewed-by: Nitin Sehgal <nsehgal@nvidia.com>