arm64: tegra: pm: use ioremap_wc
ioremap_wc maps to what ARMv8 calls Normal_NC and ioremap maps to
Device_nGnRnE.
Device_nGnRnE requires:
1. Reads do not hit in the cache.
2. Writes are written out to the bus before a memory barrier executes
3. No speculative reads
4. No combining reads or writes into a single transaction
5. No reordering reads or writes
6. No early acknowledgement from the bus. Instead wait for the device
on the bus to send an acknowledgement
Normal_NC requires:
1. Reads to not hit in the cache
2. Writes are written out to the bus before a memory barrier executes
The only requirement for the LP0 exit code is that it is written out to
memory instead of being held in cache. So, use ioremap_wc.
Bug 1512416
Change-Id: I4ab33621daa2207490ca17196d2f8805248e76ef
Signed-off-by: Alex Van Brunt <avanbrunt@nvidia.com>
Signed-off-by: Rich Wiley <rwiley@nvidia.com>
Reviewed-on: http://git-master/r/407803
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Peng Du <pdu@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>