arm: tegra: use ioremap_wc for the framebuffer
ioremap_wc maps to what ARMv8 calls Normal_NC and ioremap maps to
Device_nGnRnE.
Device_nGnRnE requires:
1. Reads do not hit in the cache.
2. Writes are written out to the bus before a memory barrier executes
3. No speculative reads
4. No combining reads or writes into a single transaction
5. No reordering reads or writes
6. No early acknowledgement from the bus. Instead wait for the device
on the bus to send an acknowledgement
Normal_NC requires:
1. Reads to not hit in the cache
2. Writes are written out to the bus before a memory barrier executes
The only requirement for the framebuffer is that it is written out to
memory instead of being held in cache. So, use ioremap_wc.
Bug 1512416
Change-Id: Ie655ff2c5c0880a59df61307830643f534f189c7
Signed-off-by: Alex Van Brunt <avanbrunt@nvidia.com>
Signed-off-by: Rich Wiley <rwiley@nvidia.com>
Reviewed-on: http://git-master/r/412777
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Krishna Reddy <vdumpa@nvidia.com>